Prevent access of a potentially freed xio instance
This commit introduces a fix for issue ARMmbed/mbed-os-example-for-azure#33. With a long term connection using SAS tokens, eventually the SAS token would expire and the IoTHub client using this MQTT client  would automatically disconnect. Currently, the only way to refresh the SAS token from the IoTHub client is to disconnect and reconnect.

During the disconnection process, the XIO transport handle being used by the MQTT client is deinitialized and freed, leaving the MQTT client with a dangling pointer to a deleted XIO transport.

This would subsequently cause hard faults on the ARM Cortex-M target.

The fix in this commit is to clear the XIO transport pointer (ie: set it to NULL) by adding a call to `close_connection`. Since the pointer is now NULL, existing checks will prevent it from being dereferenced and causing hard faults.